SequentialCompression

Documentation for SequentialCompression.

SequentialCompression.CompressedArraySeq

SequentialCompression.CompressedArraySeq

Base.append!

Base.getindex

Base.getindex

Base.ndims

Base.size

SequentialCompression.CompressedArraySeq — Type

CompressedArraySeq{T,Nx}

A mutable structure for storing time-dependent arrays in a compressed format.

Fields

data::Vector{UInt8}: Compressed data in byte form.

headpositions::Vector{Int64}: Positions of the beginning of each time slice in data.

tailpositions::Vector{Int64}: Positions of the end of each time slice in data.

spacedim::NTuple{Nx,Int32}: Dimensions of the spatial grid.

timedim::Int32: Number of time steps.

eltype::Type{T}: Element type of the uncompressed array.

tol::Float32: Mean absolute error that is tolerated.

precision::Float32: Controls the precision, bounding a weak relative error.

rate::Int64: Fixes the bits used per value.

Example

julia> using SequentialCompression

julia> compArray = CompressedArraySeq(Float64, 4, 4)
CompressedArraySeq{Float64, 2}(UInt8[], [0], [0], (4, 4), 0, Float64, 0.0f0, 0.0f0, 0)

julia> compArray.timedim
0

SequentialCompression.CompressedArraySeq — Method

CompressedArraySeq(array::AbstractArray{<:AbstractFloat}; rate::Int=0, tol::Real=0, precision::Real=0)

Create a new CompressedArraySeq from an uncompressed array.

Arguments

array::AbstractArray{<:AbstractFloat}: Initial array to compress and store.
rate::Int: [Fixes the bits used per value](https://zfp.readthedocs.io/en/release0.5.5/modes.html#fixed-rate-mode).
tol::Real: [Mean absolute error that is tolerated](https://zfp.readthedocs.io/en/release0.5.5/modes.html#fixed-accuracy-mode).
precision::Real: [Controls the precision, bounding a weak relative error](https://zfp.readthedocs.io/en/release0.5.5/modes.html#fixed-precision-mode).

Example

julia> using SequentialCompression

julia> array = [Float64(i + j) for i=1:4, j=1:4];

julia> compArray = CompressedArraySeq(array);

julia> compArray[1] == array
true

julia> compArray.timedim
1

Base.append! — Method

append!(compArray::CompressedArraySeq{T,N}, array::AbstractArray{T,N})

Append a new time slice to compArray, compressing array in the process.

Arguments

compArray::CompressedArraySeq{T,N}: Existing compressed array.
array::AbstractArray{T,N}: Uncompressed array to append.

Example

julia> using SequentialCompression

julia> compArray = CompressedArraySeq(Float64, 4, 4);

julia> append!(compArray, ones(4, 4));

julia> compArray[1]
4×4 Matrix{Float64}:
 1.0  1.0  1.0  1.0
 1.0  1.0  1.0  1.0
 1.0  1.0  1.0  1.0
 1.0  1.0  1.0  1.0

julia> compArray.timedim
1

Base.getindex — Method

getindex(compArray::CompressedArraySeq, timeidx::Colon)

Retrieve and decompress all time slices from compArray.

Base.getindex — Method

getindex(compArray::CompressedArraySeq, timeidx::Int)

Retrieve and decompress a single time slice from compArray at timeidx.

Base.ndims — Method

ndims(compArray::CompressedArraySeq)

Returns the number of dimensions of the uncompressed array, including the time dimension.

Base.size — Method

size(compArray::CompressedArraySeq)

Returns the dimensions of the uncompressed array, with the last dimension being the time dimension.